pyproject.toml
[project]
name = "monthlyspecs-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-allauth>=65.3.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "markdown>=3.9",
]
